Move user schemas closer to sql adapter
