+++ app/db/registry.py
@@ -2,4 +2,4 @@ from app.bill.schemas import Bill  # noqa:F401
 from app.order.schemas import Order, OrderItem  # noqa:F401
 from app.payment.schemas import Payment  # noqa:F401
 from app.product.adapters.sql import Product, ProductVariant  # noqa:F401
-from app.user.schemas import User, UserCredentials  # noqa:F401
+from app.user.adapters.sql import User, UserCredentials  # noqa:F401
